ListaDeReproduccion Endpoints

+++ demo/src/main/java/com/example/demo/user/infrastructure/UserRepository.java
@@ -4,6 +4,6 @@ import com.example.demo.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository<T extends User> extends JpaRepository<T, Long>{
+public interface UserRepository<T extends User> extends JpaRepository<T, Integer>{
     Optional<T> findByEmail(String email);
 }
